cursor_vector handles newline lines + file structure
