+++ includes/vec2i.h
@@ -1,0 +1,12 @@
+#ifndef VEC2I_H
+# define VEC2I_H
+
+typedef struct			s_vec2i
+{
+	int					x;
+	int					y;
+}						t_vec2i;
+
+t_vec2i					vec2i_sub(t_vec2i a, t_vec2i b);
+
+#endif
